MAINT: Refactor account component
