+++ src/model/User.tsx
@@ -1,0 +1,5 @@
+export interface UserStates {
+  username: string
+  email: string
+  created: string
+}
